+++ bootstrap3/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '5.0.3'
+__version__ = '5.1.0'